requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
crispy-tailwind==0.5.0
cryptography==36.0.2
defusedxml==0.7.1
Django==4.0.3
django-allauth==0.50.0
django-crispy-forms==1.14.0
djangorestframework==3.13.1
idna==3.3
oauthlib==3.2.0
Pillow==9.0.1
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
sorl-thumbnail==12.8.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
